An ABNF production for URI has been introduced to correspond to the common usage of the term: an absolute URI with optional fragment.
Section 2.2 on reserved characters has been rewritten to clearly explain
what characters are reserved, when they are reserved, and why they are
reserved even when not used as delimiters by the generic syntax.
Likewise, the section on escaped characters has been rewritten, and
URI normalizers are now given license to unescape any octets
corresponding to unreserved characters.  The crosshatch ("#") character
has been moved back from the excluded delims to the reserved set.

I am in the midst of rewriting the ABNF for URI to make it more
friendly to LALR parsers and significantly reduce complexity.
As a result, the layout form of syntax description has been removed,
along with the opaque-part and uric-no-slash ABNF productions.
All references to "opaque" URIs have been replaced with a better
description of how the path component may be opaque to hierarchy.
The fragment identifier has been moved back into the URI components,
though I am still working on the text.

The ambiguity regarding the parsing of net-path, abs-path, and rel-path
is now fully explained and disambiguated in the same section.

Fixed 038-qualified: The ABNF of qualified has been simplified to remove
a parsing ambiguity without changing the allowed syntax.  The toplabel
production has been removed because it served no useful purpose.
The ambiguity regarding the parsing of host as IPv4address or hostname
is now fully explained and disambiguated in the same section.


git-svn-id: https://svn.apache.org/repos/asf/labs/webarch/uri/trunk@503015 13f79535-47bb-0310-9956-ffa450edef68